requirements.txt
faiss-cpu
sentence_transformers
langchain
openai
streamlit
ai21
langchain_openai
langchain-community
scikit-learn

pdfminer
pdfminer.six
python-docx
pypdf2